requirements.txt
pywireguard
ruamel.yaml
dominate